requirements.txt
# Logging (如果使用额外的日志处理库)
python-json-logger==2.0.7

# Binance
binance-connector
binance-futures-connector

# Okx
python-okx===0.3.3

# mysql
mysql-connector-python

# daemon
python-daemon==3.1.0